created backend folder, cleaned up eye_tracker file
