+++ backend/detection_component.py
